New model: Hand Infer
Hand Infer is an RNN network for inference of players' hands. This network takes the discard and exposed tiles as input and outputs the predicted hand sequence. The dataset comes from complete gameplay screenshots. It is first processed through Tile Classifier for recognition and encoding, and then converted into data required for Hand Infer. This commit is only a template for Hand Infer and training. The trained model is saved as "/data/model/model_hand_infer.pt".

TODO: add distinguish to exposed tile sequences as pre-knowledge input; further fine-tuning of network and parameters, maybe a complete design of RNN is required.