feat: add new model HandInfer
Model HandInfer is designed to be an RNN for inference of players' hands. The network takes the discard and exposed tiles as inputs and outputs the predicted hand sequences. The dataset comes from complete gameplay screenshots. It is first processed through model TC for recognition and encoding, and then converted into data required for model HI.